repos/Flask-Scaffold-master/requirements.txt
alembic==1.0.0
aniso8601==3.0.2
autopep8==1.2.1
blinker==1.4
click==6.7
Flask==0.10.1
Flask-Cors==3.0.6
Flask-Mail==0.9.1
Flask-Migrate==1.4.0
Flask-RESTful==0.3.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
inflect==0.2.5
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.3.0
marshmallow-jsonapi==0.3.0
pep8==1.7.1
psycopg2==2.7
pycrypto==2.6.1
PyJWT==1.4.0
PyMySQL==0.6.6
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
PyYAML==3.11
selenium==3.141.0
six==1.11.0
SQLAlchemy==1.2.11
urllib3==1.24.1
Werkzeug==0.14.1
